+++ lib/aspera/cli/plugins/config.rb
@@ -89,7 +89,7 @@ module Aspera
         option :plugin_folder,      description: 'Folder where to find additional plugins',                           handler: :add_plugin_folders
         option :sdk_url,            description: 'Ascp: URL to get Aspera Transfer Executables',                      default: SpecialValues::DEF
         option :locations_url,      description: 'Ascp: URL to get download locations of Aspera Transfer Daemon',    default: Ascp::Installation.instance.transferd_urls, handler: Ascp::Installation.instance.method(:transferd_urls=)
-        option :sdk_folder,         description: 'Ascp: Path to folder with ascp (or product with "product:")',      handler: Products::Transferd.method(:sdk_directory=)
+        option :sdk_folder,         description: 'Ascp: Path to folder with ascp (or product with "product:")',      handler: Ascp::Installation.instance.method(:sdk_folder=)
         option :smtp,               schema: Schema::Registry::SMTP_OPTIONS
         option :notify_to,          description: 'Email: Recipient for notification of transfers'
         option :notify_template,    description: 'Email: ERB template for notification of transfers'